# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AJKMart Super App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpRequest"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOtpResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and login/register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/send-email-otp:
    post:
      operationId: sendEmailOtp
      tags: [auth]
      summary: Send OTP to email address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        "200":
          description: Email OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOtpResponse"
  /auth/verify-email-otp:
    post:
      operationId: verifyEmailOtp
      tags: [auth]
      summary: Verify email OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                otp:
                  type: string
              required: [email, otp]
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
components:
  schemas:
...
    SendOtpRequest:
      type: object
      properties:
        phone:
          type: string
        role:
          type: string
          enum: [customer, rider, vendor]
        preferredChannel:
          type: string
          enum: [whatsapp, sms, email]
      required: [phone]
    SendOtpResponse:
      type: object
      properties:
        message:
          type: string
        otp:
          type: string
          description: Dev only - OTP for testing
      required: [message]
    VerifyOtpRequest:
      type: object
      properties:
        phone:
          type: string
        otp:
          type: string
        role:
          type: string
          enum: [customer, rider, vendor]
      required: [phone, otp]
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    # ── User schemas ──────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
        phone:
          type: string
        name:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        avatar:
          type: string
        walletBalance:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, phone, roles, walletBalance, isActive, createdAt]
...